third_party/blink/web_tests/external/wpt/custom-elements/registries/Element-customElementRegistry-expected.txt

This is a testharness.js-based test.
[FAIL] customElementRegistry on a newly constrcuted element should return window.customElements by default
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] customElementRegistry on an element inside a declarative shadow DOM should return window.customElements by default
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] customElementRegistry on a clone of a declarative shadow tree with shadowrootcustomelementregistry should return the global registry after getting inserted into a document
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] customElementRegistry on an element inside a declarative shadow DOM with shadowrootcustomelementregistry should return the scoped registry after calling initialize
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] customElementRegistry on a builtin element created by calling createElement on CustomElementRegistry should return the registry
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] customElementRegistry on an upgarde candidate created by calling createElement on CustomElementRegistry should return the registry
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] customElementRegistry on an unknown element created by calling createElement on CustomElementRegistry should return the registry
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] customElementRegistry on a defined custom element created by calling createElement on CustomElementRegistry should return the registry
  assert_equals: expected object "[object CustomElementRegistry]" but got null
Harness: the test ran to completion.
